% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.svySE_cfg}
\alias{print.svySE_cfg}
\title{Print svySE configuration}
\usage{
\method{print}{svySE_cfg}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{"svySE_cfg"}.}

\item{...}{Argumentos adicionales.}
}
\value{
Invisiblemente el objeto \code{x}.
}
\description{
Imprime la configuracion de \code{svySE}.
Prints the \code{svySE} configuration.
}
