% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{print.svySE_result}
\alias{print.svySE_result}
\title{Print svySE result}
\usage{
\method{print}{svySE_result}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{"svySE_result"}.}

\item{...}{Argumentos adicionales.}
}
\value{
Invisiblemente \code{x}.
}
\description{
Imprime un resumen del resultado de \code{svySE_calc()}.
}
\details{
Prints a summary of the result from \code{svySE_calc()}.
}
