% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{print.svySE_simple_result}
\alias{print.svySE_simple_result}
\title{Print an svySE simple result}
\usage{
\method{print}{svySE_simple_result}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{"svySE_simple_result"}.}

\item{...}{Argumentos adicionales.}
}
\value{
Invisiblemente \code{x}.
}
\description{
Print an svySE simple result
}
