% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{svySE_chk_bool}
\alias{svySE_chk_bool}
\title{Check logical scalar}
\usage{
svySE_chk_bool(x, arg)
}
\arguments{
\item{x}{Objeto a validar / Object to validate.}

\item{arg}{Nombre del argumento / Argument name.}
}
\value{
Invisiblemente \code{TRUE}.
}
\description{
Valida que un argumento sea logico de longitud uno.
Checks whether an argument is a single logical value.
}
\keyword{internal}
