% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnas.R
\name{svySE_cols_tab}
\alias{svySE_cols_tab}
\title{Select simple table columns}
\usage{
svySE_cols_tab(type = "full", cols = NULL)
}
\arguments{
\item{type}{Tipo de salida / Output type. Opciones:
\code{"full"}, \code{"unweighted"}, \code{"target"}, \code{"freq"}, \code{"pct"},
\code{"expanded"}, \code{"expanded_freq"}, \code{"expanded_pct"}, \code{"counts"}, \code{"percentages"} o \code{"custom"}.}

\item{cols}{Vector de columnas cuando \code{type = "custom"}.}
}
\value{
Vector de columnas seleccionadas / Selected column vector.
}
\description{
Selecciona columnas para exportar la tabla simple.
Selects columns to export the simple table.
}
\examples{
svySE_cols_tab("full")
svySE_cols_tab("target")
svySE_cols_tab("expanded")
svySE_cols_tab("expanded_freq")
svySE_cols_tab("expanded_pct")
svySE_cols_tab("counts")
svySE_cols_tab("percentages")
svySE_cols_tab(
  "custom",
  cols = c("freq_1", "exp_1", "freq_total", "exp_total")
)

}
