% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnas.R
\name{svySE_cols_tab_all}
\alias{svySE_cols_tab_all}
\title{Available simple table columns}
\usage{
svySE_cols_tab_all()
}
\value{
Vector de nombres de columnas / Character vector of column names.
}
\description{
Devuelve todas las columnas disponibles para la tabla simple.
Returns all available columns for the simple table.
}
