% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{svySE_calc}
\alias{svySE_calc}
\title{Calculate sampling errors}
\usage{
svySE_calc(
  data,
  indicators,
  group_vars,
  group_labels = group_vars,
  strata,
  weight,
  division = NULL,
  div_weight = NULL,
  cfg = svySE_cfg(),
  verbose = TRUE,
  strict = FALSE
)
}
\arguments{
\item{data}{Base de datos ya cargada en R / Data frame already loaded in R.}

\item{indicators}{Vector con los nombres de los indicadores.}

\item{group_vars}{Variables de agrupacion.}

\item{group_labels}{Etiquetas de las variables de agrupacion.}

\item{strata}{Variable o variables de estrato.}

\item{weight}{Variable de peso principal.}

\item{division}{Variable de division opcional.}

\item{div_weight}{Variable de peso opcional para las divisiones.}

\item{cfg}{Configuracion creada con \code{svySE_cfg()}.}

\item{verbose}{Si es \code{TRUE}, imprime avance del calculo.}

\item{strict}{Si es \code{TRUE}, detiene el calculo cuando encuentra valores
fuera de \code{cfg$valid_values}; si es \code{FALSE}, los excluye con advertencia.}
}
\value{
Objeto de clase \code{"svySE_result"}.
}
\description{
Calcula errores muestrales para uno o mas indicadores usando un data.frame
previamente cargado por el usuario.
}
\details{
Calculates sampling errors for one or more indicators using a data.frame
previously loaded by the user.
}
