% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnas.R
\name{svySE_cols_err}
\alias{svySE_cols_err}
\title{Select sampling error columns}
\usage{
svySE_cols_err(type = "full", cols = NULL)
}
\arguments{
\item{type}{Tipo de salida / Output type. Opciones:
\code{"full"}, \code{"pct"}, \code{"abs"}, \code{"basic"}, \code{"quality"} o \code{"custom"}.}

\item{cols}{Vector de columnas cuando \code{type = "custom"}.}
}
\value{
Vector de columnas seleccionadas / Selected column vector.
}
\description{
Selecciona columnas para exportar la tabla de errores muestrales.
Selects columns to export the sampling error table.
}
\examples{
svySE_cols_err("full")
svySE_cols_err("pct")
svySE_cols_err("custom", cols = c("est_pct", "cv"))

}
