% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnas.R
\name{svySE_cols_err_all}
\alias{svySE_cols_err_all}
\title{Available sampling error columns}
\usage{
svySE_cols_err_all()
}
\value{
Vector de nombres de columnas / Character vector of column names.
}
\description{
Devuelve todas las columnas metricas disponibles para la tabla de errores
muestrales.
Returns all available metric columns for the sampling error table.
}
