% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnas.R
\name{svySE_cols_tab}
\alias{svySE_cols_tab}
\title{Select simple table columns}
\usage{
svySE_cols_tab(type = "full", cols = NULL)
}
\arguments{
\item{type}{Tipo de salida / Output type. Opciones:
\code{"full"}, \code{"target"}, \code{"freq"}, \code{"pct"} o \code{"custom"}.}

\item{cols}{Vector de columnas cuando \code{type = "custom"}.}
}
\value{
Vector de columnas seleccionadas / Selected column vector.
}
\description{
Selecciona columnas para exportar la tabla simple.
Selects columns to export the simple table.
}
\examples{
svySE_cols_tab("full")
svySE_cols_tab("target")
svySE_cols_tab("custom", cols = c("freq_1", "pct_1"))

}
