% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{svySE_is_cfg}
\alias{svySE_is_cfg}
\title{Check svySE configuration object}
\usage{
svySE_is_cfg(x)
}
\arguments{
\item{x}{Objeto a evaluar / Object to check.}
}
\value{
\code{TRUE} o \code{FALSE}.
}
\description{
Verifica si un objeto pertenece a la clase \code{svySE_cfg}.
Checks whether an object belongs to class \code{svySE_cfg}.
}
\keyword{internal}
