% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar.R
\name{svySE_xlsx}
\alias{svySE_xlsx}
\title{Export svySE results to Excel}
\usage{
svySE_xlsx(
  x,
  file_err = "svySE_errors.xlsx",
  file_tab = "svySE_simple.xlsx",
  cols_err = svySE_cols_err("full"),
  cols_tab = svySE_cols_tab("full"),
  start_row = 10,
  overwrite = TRUE,
  keep_na = FALSE
)
}
\arguments{
\item{x}{Objeto generado por \code{svySE_calc()}.}

\item{file_err}{Archivo Excel para errores muestrales.}

\item{file_tab}{Archivo Excel para tabla simple.}

\item{cols_err}{Columnas de error a exportar.}

\item{cols_tab}{Columnas de tabla simple a exportar.}

\item{start_row}{Fila inicial de escritura.}

\item{overwrite}{Sobrescribir archivos existentes.}

\item{keep_na}{Mantener NA en Excel.}
}
\value{
Invisiblemente \code{TRUE}.
}
\description{
Exporta los resultados generados por \code{svySE_calc()} a archivos Excel.
Exports results generated by \code{svySE_calc()} to Excel files.
}
